Refactor StationMetadata into several modules
